expansion on the release_initializer.js shim for multiple scheme snippets
details: status quo ante, only the body of the first <script> tag for biwascheme{,-min}.js would be executed.

changes: now execute a) the body of *every* <script> tag to a file with a name containing "biwascheme" (this may need rethinking), as well as b) the src (if any) of every <script> tag with the class "biwascheme", or the body otherwise.